+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
- * Bumped runtimeVersion to `2022.02.03`.
+ * Bumped runtimeVersion to `2022.02.09`.
+ * Upgrade stable Dart analysis SDK to `2.16.1`.
+ * Upgraded preview Dart analysis SDK to `2.16.1`.
  * Upgraded stable Flutter analysis SDK to `2.10.0`.
  * Upgraded preview Flutter analysis SDK to `2.10.0`.